# Improove performance colors