--- a/src/ftxui/screen/color.cpp
+++ b/src/ftxui/screen/color.cpp
@@ -35,8 +35,7 @@ const std::array<const char*, 33> palette16code = {
 }  // namespace
 
 bool Color::operator==(const Color& rhs) const {
-  return red_ == rhs.red_ && green_ == rhs.green_ && blue_ == rhs.blue_ &&
-         type_ == rhs.type_;
+  return color.all == rhs.color.all && type_ == rhs.type_;
 }
 
 bool Color::operator!=(const Color& rhs) const {
@@ -49,17 +48,17 @@ std::string Color::Print(bool is_background_color) const {
       return is_background_color ? "49"s : "39"s;
 
     case ColorType::Palette16:
-      return palette16code[2 * red_ + is_background_color];  // NOLINT;
+      return palette16code[2 * color.channel.red_ + is_background_color];  // NOLINT;
 
     case ColorType::Palette256:
-      return (is_background_color ? "48;5;"s : "38;5;"s) + std::to_string(red_);
+      return (is_background_color ? "48;5;"s : "38;5;"s) + std::to_string(color.channel.red_);
 
     case ColorType::TrueColor:
     default:
       return (is_background_color ? "48;2;"s : "38;2;"s)  //
-             + std::to_string(red_) + ";"                 //
-             + std::to_string(green_) + ";"               //
-             + std::to_string(blue_);                     //
+             + std::to_string(color.channel.red_) + ";"                 //
+             + std::to_string(color.channel.green_) + ";"               //
+             + std::to_string(color.channel.blue_);                     //
   }
 }
 
@@ -73,16 +72,18 @@ Color::Color(Palette1 /*value*/) : Color() {}
 
 /// @brief Build a transparent using Palette16 colors.
 /// @ingroup screen
-Color::Color(Palette16 index) : type_(ColorType::Palette16), red_(index) {}
+Color::Color(Palette16 index) : type_(ColorType::Palette16) {
+    color.channel.red_ = index;
+}
 
 /// @brief Build a transparent using Palette256 colors.
 /// @ingroup screen
-Color::Color(Palette256 index) : type_(ColorType::Palette256), red_(index) {
+Color::Color(Palette256 index) : type_(ColorType::Palette256), color(index) {
   if (Terminal::ColorSupport() >= Terminal::Color::Palette256) {
     return;
   }
   type_ = ColorType::Palette16;
-  red_ = GetColorInfo(Color::Palette256(red_)).index_16;
+  color.channel.red_ = GetColorInfo(Color::Palette256(color.channel.red_)).index_16;
 }
 
 /// @brief Build a Color from its RGB representation.
@@ -93,7 +94,7 @@ Color::Color(Palette256 index) : type_(ColorType::Palette256), red_(index) {
 /// @param blue The quantity of blue [0,255]
 /// @ingroup screen
 Color::Color(uint8_t red, uint8_t green, uint8_t blue)
-    : type_(ColorType::TrueColor), red_(red), green_(green), blue_(blue) {
+    : type_(ColorType::TrueColor), color(red, green, blue) {
   if (Terminal::ColorSupport() == Terminal::Color::TrueColor) {
     return;
   }
@@ -118,10 +119,10 @@ Color::Color(uint8_t red, uint8_t green, uint8_t blue)
 
   if (Terminal::ColorSupport() == Terminal::Color::Palette256) {
     type_ = ColorType::Palette256;
-    red_ = best;
+    color.channel.red_ = best;
   } else {
     type_ = ColorType::Palette16;
-    red_ = GetColorInfo(Color::Palette256(best)).index_16;
+    color.channel.red_ = GetColorInfo(Color::Palette256(best)).index_16;
   }
 }
 
@@ -188,7 +189,7 @@ Color Color::Interpolate(float t, const Color& a, const Color& b) {
       }
 
       case ColorType::Palette16: {
-        const ColorInfo info = GetColorInfo(Color::Palette16(color.red_));
+        const ColorInfo info = GetColorInfo(Color::Palette16(color.color.channel.red_));
         *red = info.red;
         *green = info.green;
         *blue = info.blue;
@@ -196,7 +197,7 @@ Color Color::Interpolate(float t, const Color& a, const Color& b) {
       }
 
       case ColorType::Palette256: {
-        const ColorInfo info = GetColorInfo(Color::Palette256(color.red_));
+        const ColorInfo info = GetColorInfo(Color::Palette256(color.color.channel.red_));
         *red = info.red;
         *green = info.green;
         *blue = info.blue;
@@ -205,9 +206,9 @@ Color Color::Interpolate(float t, const Color& a, const Color& b) {
 
       case ColorType::TrueColor:
       default: {
-        *red = color.red_;
-        *green = color.green_;
-        *blue = color.blue_;
+        *red = color.color.channel.red_;
+        *green = color.color.channel.green_;
+        *blue = color.color.channel.blue_;
         return;
       }
     }