@clement-roblot 2023-08-05
Indeed, my changes did break stuff. Looking deeper into it, I'm changing a lot of stuff. It makes me feel like the performance improvement is not worth such bug changes, and we should just close the PR. Do you agree @ArthurSonzogni?